.gitignore - Add .ant-targets-build.xml file to gitignore.
I see that an ".ant-targets-build.xml" file was occasionally committed in 3c3cf997d48601ad3124cc55a2f80c89c8dbfa14 in 7.0.x branch. I think that it does not belong to the source code and that it is better to ignore this file.

By the way, ".ant-targets-build.xml" was a plain text file and not a valid XML one.